requirements.txt
langchain
wikipedia
langchain-community
langchain-openapi
langchain-anthropic
langchain-groq
python-dotenv
pydantic
duckduckgo-search
ddgs